% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_ClinicalOutcome.R
\name{Oncomine_ClinicalOutcome}
\alias{Oncomine_ClinicalOutcome}
\title{To clear Oncomine data for Clinical Outcome data}
\usage{
Oncomine_ClinicalOutcome(filenames)
}
\arguments{
\item{filenames}{The filenames that your data stored in. if ONLY data files in your working directory this object can be omitted.}
}
\value{
a dataframe
}
\description{
To clear Oncomine data for Clinical Outcome data
}
\examples{
Oncomine_ClinicalOutcome()
}
