% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_getHeaders.R
\name{Oncomine_getHeaders}
\alias{Oncomine_getHeaders}
\title{To get column headers}
\usage{
Oncomine_getHeaders(data)
}
\arguments{
\item{data}{the data name}
}
\value{
string
}
\description{
To get column headers
}
\examples{
Oncomine_getHeaders(n)
}
