% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Oncomine_heatmap_Comparison.of.All.Genes.in.CellLine.R
\name{Oncomine_heatmap_Comparison.of.All.Genes.in.CellLine}
\alias{Oncomine_heatmap_Comparison.of.All.Genes.in.CellLine}
\title{To get data from All Genes in CellLine}
\usage{
Oncomine_heatmap_Comparison.of.All.Genes.in.CellLine(filename)
}
\arguments{
\item{filename}{filename can be one or more, or missing if only data files in your working directory.}
}
\value{
a dataframe
}
\description{
To get data from All Genes in CellLine
}
\examples{
Oncomine_heatmap_Comparison.of.All.Genes.in.CellLine()
}
