% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_Metastatic.R
\name{Oncomine_Metastatic}
\alias{Oncomine_Metastatic}
\title{To clear Oncomine data for Metastatic Event Status}
\usage{
Oncomine_Metastatic(filenames, NoValue = TRUE)
}
\arguments{
\item{filenames}{The filenames that your data stored in. if ONLY data files in your working directory this object can be omitted.}

\item{NoValue}{a logistic value}
}
\value{
a dataframe
}
\description{
To clear Oncomine data for Metastatic Event Status
}
\examples{
Oncomine_Metastatic()
}
