% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_Recurrence.R
\name{Oncomine_Recurrence}
\alias{Oncomine_Recurrence}
\title{To clear Oncomine data for Recurrence Status}
\usage{
Oncomine_Recurrence(filenames, NoValue = TRUE)
}
\arguments{
\item{filenames}{The filenames that your data stored in. if ONLY data files in your working directory this object can be omitted.}

\item{NoValue}{a logistic value}
}
\value{
a dataframe
}
\description{
To clear Oncomine data for Recurrence Status
}
\examples{
Oncomine_Recurrence()
}
