% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_SampleFilters.R
\name{Oncomine_SampleFilters}
\alias{Oncomine_SampleFilters}
\title{To clear Oncomine data for Sample Filters data}
\usage{
Oncomine_SampleFilters(filenames, NoValue = TRUE)
}
\arguments{
\item{filenames}{The filenames that your data stored in. if ONLY data files in your working directory this object can be omitted.}

\item{NoValue}{a logistic value}
}
\value{
a dataframe
}
\description{
To clear Oncomine data for Sample Filters data
}
\examples{
Oncomine_SampleFilters()
}
