% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_SampleRecurrence.R
\name{Oncomine_SampleRecurrence}
\alias{Oncomine_SampleRecurrence}
\title{To clear Oncomine data for Sample Recurrence Status}
\usage{
Oncomine_SampleRecurrence(filenames, NoValue = TRUE)
}
\arguments{
\item{filenames}{The filenames that your data stored in. if ONLY data files in your working directory this object can be omitted.}

\item{NoValue}{a logistic value}
}
\value{
a dataframe
}
\description{
To clear Oncomine data for Sample Recurrence Status
}
\examples{
Oncomine_SampleRecurrence()
}
