% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_SurvAnalysis.Data.R
\name{Oncomine_SurvAnalysis.Data}
\alias{Oncomine_SurvAnalysis.Data}
\title{To do get data for survival analysis from Oncomine}
\usage{
Oncomine_SurvAnalysis.Data(data, Time, Outcome, OutcomeCode1, x)
}
\arguments{
\item{data}{data name}

\item{Time}{The complete variable name of following time}

\item{Outcome}{The complete variable name of outcome}

\item{OutcomeCode1}{The primary ending of outcome, which will be coded as 1. String is needed}

\item{x}{Complete Expression Value variable name. String is needed. Please use original Expression Value, this function will divide it into 1,as high expression and 0, as low expression.}
}
\value{
a Dataframe
}
\description{
Please use original Expression Value, this function will divide it into 1,as high expression and 0, as low expression.
}
\examples{
Oncomine_SurvAnalysis.Data(data=data,Time="Time",Outcome="Outcome",OutcomeCode1="Dead",x="Expression Value")
}
\author{
Jing Zhang
}
