% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_SurvAnalysis.DataCombine.R
\name{Oncomine_SurvAnalysis.DataCombine}
\alias{Oncomine_SurvAnalysis.DataCombine}
\title{To combine two datasets following command Oncomine_Survival.Data}
\usage{
Oncomine_SurvAnalysis.DataCombine(data1, data2)
}
\arguments{
\item{data1}{data1}

\item{data2}{data2}
}
\value{
a data frame
}
\description{
To combine two datasets following command Oncomine_Survival.Data
}
\examples{
Oncomine_SurvAnalysis.DataCombine(data1,data2)
}
