% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_SurvAnalysis.Plot.R
\name{Oncomine_SurvAnalysis.Plot}
\alias{Oncomine_SurvAnalysis.Plot}
\title{To plot Survival Curve after command Oncomine_SurvAnalysis.Data}
\usage{
Oncomine_SurvAnalysis.Plot(data, direction = "low")
}
\arguments{
\item{data}{data name}

\item{direction}{the curve direction,"low" or "l" means pct methord as defaulted, "high" or "h" means cumhaz methord, "event" or "e" means cumhaz methord in command ggsurvplot}
}
\description{
To plot Survival Curve
}
\examples{
Oncomine_SurvAnalysis.Plot(data)
}
\author{
Jing Zhang
}
