% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_Survival.R
\name{Oncomine_Survival}
\alias{Oncomine_Survival}
\title{To clear Oncomine data for survival data}
\usage{
Oncomine_Survival(filenames, NoValue = TRUE)
}
\arguments{
\item{filenames}{The filenames that your data stored in. if ONLY data files in your working directory this object can be omitted.}

\item{NoValue}{a logistic value}
}
\value{
a dataframe
}
\description{
To clear Oncomine data for survival data
}
\examples{
Oncomine_Survival()
}
