% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_ToGraphPad.R
\name{Oncomine_ToGraphPad}
\alias{Oncomine_ToGraphPad}
\title{To convert data to GraphPad Prism data, after command Oncomine_SurvAnalysis.Data}
\usage{
Oncomine_ToGraphPad(data)
}
\arguments{
\item{data}{data}
}
\value{
a data frame
}
\description{
To convert data to GraphPad Prism data, after command Oncomine_SurvAnalysis.Data
}
\examples{
Oncomine_ToGraphPad(data)
}
