% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_TreatmentResponse.R
\name{Oncomine_TreatmentResponse}
\alias{Oncomine_TreatmentResponse}
\title{To clear Oncomine data for Patient Treatment Response}
\usage{
Oncomine_TreatmentResponse(filenames, NoValue = TRUE)
}
\arguments{
\item{filenames}{The filenames that your data stored in. if ONLY data files in your working directory this object can be omitted.}

\item{NoValue}{a logistic value}
}
\value{
a dataframe
}
\description{
To clear Oncomine data for Patient Treatment Response
}
\examples{
Oncomine_TreatmentResponse()
}
